Almost there
Tested things without saving the password. The prompt pops, and it even
prompts twice if you typo the password. I meant to test that :)

A few minor changes to the logic. Deselecting the save password doesn't
require testing the connections, so that had to be reworked. I would like to
make it more obvious that the test connection button has been disabled.

I *think* the db merge just works. I was at least able to do the whole
populating child tables update, which makes me happy. If I read the rest
properly, then it should be good. I think I can test this, so will before the
next checkin.