Making a list, once
This is getting progressively closer to a complete rewrite of
DatabaseSchemaHelper. mea culpa.

This started because I didn't like the list I had embedded in convertDatabase.
I was able to get how the field was defined from the QSqlField object, so that
should have made it easy. It didn't, because there was no definitive list of
tables in our database. Brewtarget::DBTable only had about half of them.

That started me down a long, long chain of how to do this such that it can be
maintained and makes adding tables as easy as possible. It was also important
that I be able to use this information to properly rebuild a database, which
means order is important. This checkin is about 70% of that solution.

I've added a new table to the database called bt_alltables. Each row is a name
of a table, the value of that table in Brewtarget::DBTable and three booleans
that will affect how the information gets used. The table is created, but I
still need to do the work in Database to take advantage of it. That will be
the next checkin. Nothing is broken with this checkin, but it is really only
half functional.