Big changes for no real effect
This took far more effort than I would have actually expected, all to simply
remove a few hardcoded lists.

The metatable has been reworked to actually fit how it is being used. That
accounts for (some) of the changes to DatabaseSchemaHelper. Most of the rest
are just trying to make it a bit more consistent.

Since I am adding a new table, I increased the database version. Which means I
added a new migrate_ method.

I tweaked the option dialog a little bit to make sure I don't save the
password if the upgrade fails, and to find a color for the "test connection"
button I like and can read. Bright yellow fails the second criteria.

The major changes are to database.cpp. Some lists are no longer hard coded,
which make me happy but I'm not sure it was worth the effort. It does mean you
only need to add the table into DatabaseSchemaHelper and add a new value into
the Brewtarget::DBTable enum and the rest should just work.

There's more testing to be done tomorrow. I really, really need to find
every place we write to the database and do something we've never done before
-- check return codes. Now that we are on the network, all things can fail and
we should probably at least log some of them.