SQLite works again, as does upgrading
I promised to make sure things still worked with the default system. They
didn't, but they do now.

The biggest problems were that I was using the existence of the settings table
to determine if the database was initialized. Unfortunately, the settings
table didn't exist until version 4, and we ship an earlier db. It caused
issues. I've since fixed that by assuming that an uninitialized db will have
no tables in it. This could break.

The other problem was a bit more devious. The upgrade to version 210 uses
create_childrenTables() and create_inventoryTables(), which ultimately try to
put a line into bt_alltables, which isn't created until the upgrade to version
6. I added a simple boolean to stop that behaviour. I am not certain the
bt_alltables was a good idea.

I finally "added" the methods to go from pgsql->sqlite. I realized I just
needed a bit more thought in some existing methods, a little renaming and the
same code that copied from sqlite->pgsql can do the work. It does require an
*empty* sqlite database. Not sure if I should put something in place to catch
that or not.

I did a bit of interface tweaking to make this all happen.