addIngredientToRecipe is now commit safe
addIngredientToRecipe is modifying three separate tables. If any one of those
failed, the database would be left in an inconsistent state. So I wrapped them
all in a transaction boundary. Error checking and handling has been enhanced.

I also modified Database::copy a little. The pervious incarnation created an
empty record, then performed an update. This seemed a little odd when simply
doing an insert into got the work done, and atomically at that. So I made that
work.