Database is getting some rework.
I am really trying to just keep myself to adding transaction boundaries where
it makes sense. There are a slew of methods in database.cpp that don't really
belong there -- all of the addToRecipe and removeToRecipe methods should just
be add() and remove() in the recipe object. To get the transactions and error
returns right, I pretty much have to rewrite these methods, so I am going to
do it right.

This commit compiles. I don't know that it runs. It's late and I want some
sleep, so testing will come later.