A tale of two deletes
This was interesting. To get transactions and return codes, I am slowly going
over the database interactions and trying to make sense of them.

Today, I am focused on removing rows from the database. We actually have two
completely distinct deletes. When an ingredient is deleted from a recipe is
one delete, and when an ingredient is deleted from the trees is another.

In the first case, we need to:
o remove the entry from *_in_recipe
o remove the entry from *_children
o remove the ingredient.
The reason is that we make a copy of the ingredient and attach it to the
recipe. Once that copy is removed from the recipe, it can never be accessed or
recovered. Therefore, it should just be deleted.

In the second case, we just need to flip the deleted flag. The intent has
always been to have a "view deleted" option that would show all of the things
that have been deleted from the trees.

This commit separates the two use cases and reduces a lot of redundant
code with some clever classInfo() work.

I think we really need to consider how this database works.  To use the hops
as an example, when we add a hop to a recipe, we create a copy of the original
and link it to the recipe via the hop_in_recipe table. That makes sense of
that copy is being used in multiple recipes. It isn't. I *think* we could
simplify a lot of this stuff if we just added a foreign key to the hop table.
This is pretty much how brewnotes are already working.