More templates, more error checking
I am somewhat pleased with this one. I was able to collapse a lot of the
redundant code in Database::new* without having to engage in the horrible
switch-case statement I did in remove(). I was also about to collapse the
new*() and the new*(other) methods into one.

It compiles, but I have no clue if it runs. And its too late to test.