Scale on input
Assuming I did not utterly and completely make a mess of it, this should allow
us to scale input. What I mean by that is if you specify the scale of a field
to be (for example) grams, and then type 19 into that field, brewtarget will
now interpret this as 19 grams, not 19 kg.

The unit&scale mechanisms were about 90% in place. I had to modify a few
method signatures, a few method invocations in the tables and ONE LINE in
BtLineEdit. Everything should play by the rules now.