More try-throw-catch
Slowly working my way through the addToRecipe methods. I thought about moving
them to Recipe, and I'm still thinking of it. The syntax would be cleaner --
instead of Database::instance().addToRecipe(recipe, thingy) it would be
recipe->add(thingy). Just need to figure out if I can.

But not on this commit.