Closes #117
It wasn't so much a bug as an intentional design choice. The problem I avoided
solving was that the tree didn't know which editor to call.

I fixed that by modifying the signal to call a BtTreeView slot, which figured
out the folder and called the proper editor. It required me adding a new
method to all of the editors, and modifying the signal slightly.