This may be a problem
tl;dr is don't call recalcAll inside a transactional boundary.

So. Transaction boundaries are tough, mostly because we've some really
squirrely calls going on. I would really love to untangle the calls, but that
will be a few month's effort that I don't feel like devoting.  We really suck
at return codes. I'm dumping core left, right, sideways and center because
database methods now return NULL when we never expected it. And yes, I am at
least as guilty as anybody, and likely more so. The tree models are going to
hurt.

The best example of the squirrely is setDisplay(false) anywhere in the
Database class. This calls BeerXMLElement::setDisplay, which calls
BeerXMLElement::set() which calls Database::updateElement(), which calls
SetterCommand::append() and then SetterCommand::redo(), which calls
SetterCommand::setterStatements() before writing to the database. In this
example, it is almost impossible to figure out where the transaction
boundaries are, who should open them and when.

It is also somewhat ...  useless. The point of the SetterCommand class is to
provide undo (which I still don't think can be made to work). If we are
importing something, or copying something or ... we don't need to push
modifying the display onto the undo stack -- deleting the created element is
enough.

The other example is something like Recipe::recalcAll(), which we use an
obscene number of times everywhere. This isn't really bad, but I am not really
sure the Database class should be doing that.  recalcAll() will save the
recalculated values to the database, which calls Recipe::set, which calls
BeerXMLElement::set()... and wreaks all sorts of havoc with the transaction
boundaries. And do we really have to? We emit a bazillion signals that should
cause the right thing to happen without invoking the overhead that is
recalcAll.