Time to bury SetterCommand
Sometimes, cleaning things up feels better than it should. SetterCommand is no
more. In consulting with rocketman, we both felt that what SetterCommand
original intended to do -- provide an undo stack -- simply wasn't possible in
its current form. Instead, it just made an already fairly complex part of the
code more complex and was wreaking havoc on the transactional boundaries.

So I moved the necessary parts to Database::removeRecord and
Database::updateEntry and removed SetterCommand.

This checkin is stable and is known to work against postgresql. I still need
to take a few minutes and test against sqlite. My testing is to: copy a recipe
(which tests out almost every table we have), brew it and then change the boil
time.