Final commit for throwing a database
Finished making Database do try/catch. A minor change was requried in
OptionsDialog to support these changes. I also made a few methods return void
instead of bool.

I am a little less confident in the changes I made to sqlDatabase(). But it
seems to me we are opening a new connection, and rather have to supply the
connection information. Although it worked previously, so maybe we don't? It
runs, and seems to do what it should.