Small fixes and some ideas on the fly sparge
I fixed the volume problem. As I somewhat expected, it was a side effect of
disabling the signals. The fix was to add a method to figure volume lost
to the grain. There are still some oddities, but it seems to do the right
thing.

Started playing with the fly sparge settings a little. It will get more into
the brewing methods idea I have, but curently you can't add another step after
you select "fly sparge". I am going to continue working the idea, but this is
the start.