Closes 206 - core while populating child tables
populateChildrenTables() requires the tableName() hash to be populated, but
the hash can't be populated until after DatabaseSchemaHelper::migrate is
called. With the way the methods are written now, this leads to a chicken-egg
sort of problem.

I solved the problem by simply have populateChildrenTables() prepopulate the
hash. On the one time we need to call that method, the hash will get
initialized twice. I could maybe fix that, but it seemed an awful lot of work
for something that should happen once. In a broader question, shouldn't the
populateChildrenTables be part of the migrate code?

The changes to the header files are simply cleaning some compiler warnings.