Closes #211 - Crash recreating brewday instructions.
The easy part is the problem defintion. In one of my multiple refactorings,
the removeIngredientFromRecipe method attempted to remove a line from
instruction_children, and that table doesn't exist.

I thrashed around a little, but it is one exception for just one class and I
didn't feel a more generic solution was required. This is, admittedly, a
little hackish and I will likely come to regret it later.

The change to database.h is removing a qDebug() line that I swear I have
removed at least three times already.