Closes #127 - TravisCI can't find webkit
Qt 5.6 did away with the web kit, but none of our build environment have the
replacement yet. This patch replaces all the WebDocs with TextBrowser, which
seems to work and should compile.

I removed the webkit requirement from the CMakeLists file, but I wasn't sure
if the webkitwidgets needed axed as well. Although writing it out makes the
answer somewhat obvious.