Fixes #221 - Can't add mash step in dev
Two problems, and a few cleanups.

The first problem was that newMashStep() wasn't calling insertNewMashStep(),
so the extra two steps to connect the mashstep to its mash weren't being done.
In fixing this, I was able to remove both insertNewDefaultIngredient and
insertNewDefaultMash. Nice little bonus there.

The second problem was I had missed a signal when I removed the SetterCommand
class. Basically, it never signaled anything had changed. I added the signal
back in, and things are happy again. I've tested by mucking around in the
mashstep editor and playing with the hops.

In the process, I found a few methods that hadn't been updated to use the hot
new try-catch process or were doing it wrong. I cleaned those up.