Closes #230 -- buggy refactor
Testing on _section.isValid() didn't work after I defined the Q_PROPERTY. It
set the property to an empty string, which was valid if empty. Since we almost
never defined configSection on a BtLineEdit, this meant we didn't invoke the
complex math to find the proper configSection.

This commit fixes that. In the process if chasing this problem down, I found a
few dialogs that were not properly defining the configSection, so I fixed that
too.