Wonky inputs
I missed something important when I implemented the scales for input. as it
currently stands, we will *always* use the default unit/scale regardless of
what the user inputs. We should only be doing that when NO unit/scale has been
entered.

I am slowly detangling the mess, and I think I have a fairly simply solution,
but I have already made a hash of this once and am trying to take my time to
verify the solution.