Closes #244 -- New mash type are behaving poorly (for reals)
According to both sqlite and postgres, the order in which rows are returned is
undefined unless an order by clause is used. In postgres, it seems "undefined"
means random. In sqlite, it seems "undefined" means in row-id order.

When we got the mash steps, we didn't specify the order. It happened to work
for sqlite, but not for postgresql. This patch adds an order by clause to make
sure it isn't "undefined".