Closes #255 - Adjusting volume on preboil tab
Somewhat simple fix. Basically, when I did the scales for input, I changed the
symantics of what the "force" flag means in BtLineEdit. I just didn't change
everything.

There's a signal that needs to be emitted when the field changes, but I only
want it emitted when the user enters a new value in the field. The previous
logic used the forced flag, which was clever but wrong. When I changed the
symantics, I should have changed the flag. This commit fixes that.

And a new rule. Don't use one flag to do two different things. Use different
flags.