Bug/Feature #247 -- Database vanished
This doesn't fix the core problem. That comes next, because I think I can.
This implements automatic, under the cover backups.

The OptionDialog has been significantly re-arranged. The user can now select
how often (aka, frequency) to backup, and how many old copies to keep.
Frequency is defined by the number of times a user closes brewtarget and accepts
the changes. Frequency defaults to 4, the number defaults to 10. A frequency
of 1 means always backup on close. If the number to keep is -1, it means never
delete; 0 means never backup; >0 means that number of backups

When brewtarget is closed, we get a counter from the properties file. If that
counter % frequency is not 0, we simply store the updated counter and exit.

If the number to save is 0, we exit.

If we get this far, we pull the backup directory name and the list of existing
files from the properties file.

If counter % frequency is 0, we generate a filename like bt_database.yyyyMMdd.
We then check to see if that file exists in the backup directory. If it does,
we go into a loop appending 0001, 0002 ... until we get a filename that
doesn't exist. If we get to 10000, we give up and just use
bt_database.yyyyMMdd.

Once we have the name, we call backupToDir() which already existed. That just
copies the file for us. We then append the new name to the list.

The final step is to clean up the old copies if we need to. While the size of
the list is greater than the number of copies to keep, we pop the first (aka,
oldest) file name off the list, and delete it. Using a while loop allows the
user to adjust the number of preserved copies down and brewtarget will clean
up after itself.

Finally, we store the new list of files names back into the properties file
and exit.