Closes #267
That was .. a mess. I don't know how far back this error goes, but I think its
old.

First, the SQL statement in Database::swapMashStepOrder was wrong. As
@greggree noticed, it was using msid instead of mash_id. The update itself
changed the stepNumber to be the id of the steps, which ... well, just wasn't
right. So I fixed that, and reworked the QString a little.

Second, there was nothing in the MashStepTableModel to do the right thing when
the stepNumber changed signal was recieved. I added a method that just removes
the rows and re-adds them. I do not disconnect any signals, which may come to
haunt us later.

I think this fixes the issues.