Closes #272 -- Default database directory location
Late initialization for the win. Basically, QDir::canonicalPath() will return
an empty string if the directory doesn't exist. When starting with no existing
~/.config/brewtarget, this caused an empty user_data_dir string to be written
to the config and every start after that tried to find its db in cwd.

I simply delayed creating the config entry until after we ensured all the
directories existed.