Add Save/Cancel button to equipment warnings
Currently if you make a mistake in the equipment editor you are warned
about the error, but are not given a chance to correct the problem
before the change is saved.

This commit removes the current default QMessageBox::OK button from the
warning box, and replaces it with QMessageBox::Save/Cancel buttons.
This change gives a user a chance to make corrections with a minimal
number of interactions without changing context.