Closes #291 - Segfault merging database
In all of my wonderful refactors, I missed the function pointers in
Database::makeTableParams(). The function pointers resulted in newYeast (for
example) being called with 0x1 as the pointer to other, which lead to a
sigsegv.

I've modified the function pointers so they do the approximate right thing,
and no cores. I also had to clean up a small problem with an insert statement.