Lambdas ARE better than function pointers
That was fun. So to do Database::makeTableParams() with lambdas instead of
function pointers took a little more than I had expected.

To use the [&] capture, Database::makeTableParams() cannot be static mostly
because there's no context to be captured in a static method. So I had to
change that all around.  Since the resulting structure was only used by
updateDatabase(), I removed the static from the method, and called it from
updateDatabase(). That gave the required context.

Currently, the lambdas are inferring the return type based on the signature of
the return. This may come to haunt us later, but it really does make for a
nice declaration.

The hard part was figuring out the std::function piece. I started with
std::function<void()> since that was what all teh cool kids were doing on the
internet. But I kept getting compile errors saying I wasn't ignoring a void
return like I should be. It took me more thought that it should have, but the
proper signature is std::function<BeerXMLElement*()>.

The last step was to update the CMake file to include the std=c++11 flag. That
isn't default for Qt until 5.7 I think, and I'm building this on Qt5.6 so far.