Closes #282 -- batch sparge type has no affect on water slider amount range.
This was definitely part of the batch sparge type stuff. I didn't call
updateMaxAmt() when the type was changed to batch or when it was changed back
to infusion.

It also drove me nuts that the field wouldn't change until after you entered
the temp. So I fixed that too.