Closes #283 -- temperature slider reversal
The slider didn't really reverse. Its behavior simply didn't make sense.

The max temp label is taken either as the configured boiling point of water
(bpow, for short) or 100C. In other words, it is treated mostly like a
constant and not calculated.

The min temp label, on the other hand, is calculated. We never bothered to
check if the results made any sense. While I kind of see the point of staying
true to the maths, I think we should consider physics too.

I have put a number of checks in place to make sure if the temperature exceeds
the configured bpow, the temperature is set to the bpow. The min/max labels
will not exceed the bpow, the temperature label will not, the step starting or
target temps will not either.