Closes #116 - Crash at closing after deleting hop
Deleting had nothing to do with it, but adding a new hop did. For reasons I
still do NOT understand, the hop table was attempting to refresh itself on
delete. I went through any number of interations, but could never quite figure
out why.

This is the easiest of the fixes I found -- the database isn't closed until
after we tear all the windows down. Since MainWindow doesn't open the DB, it
shouldn't close it either. I still cannot explain why the hop editor alone
does this, and not the misc editor.