Closes #306 - Copy databsae doesn't copy the settings table properly
I had previously been skipping the settings table when copying a database.
That didn't work, but I couldn't use an insert line. This solution is to
special case the settings table and use an SQL UPDATE instead of an INSERT.

It required a new method, renaming an existing method and changing some
variable names to better indicate what they are doing.