Closes #309 - Database performance
If you read the issue, most of this commit should look familiar. The big
changes are explained in there.

Previously, requested the entire row for an item, and then returned our one
column. Now, we request the one column we want, for a 50% speed improvement.
This required us to change how we prepare and cache our queries. selectAll is
gone, as is the method that created it. selectSome is used now, and built on
the fly.

We now cache four attributes: name, folder, deleted and display. Eachof those
attributes is now defined in BeerXMLElement. This required removing all of the
code from equipment, fermentable, hop, instruction, mash, mashstep, misc,
recipe, style, water and yeast that dealt with setting and getting names.

These two changes resulted in the load times for brewtarget to drop from 15
seconds to 2 seconds when using postgresql, and we now load the sqlite
database in about 0.5 seconds.

While I was messing about, I also deleted some commented out code.