Properly escape SQL statements.
While initialising the database for a new user a random issue occurred.
Every now and then the initialisation failed with a SQL error.
The issue seemed to happen with a record named something like
»Harry's rolled…«. I have not investigated why it is random, nor was I
able to find a similar record in the database.

In order to reproduce the issue I modified the default_db.sqlite database.
I changed the fermentable entry for »Acid Malt« to »Acid Malt'X«. This led
to a reproducible error:

[20:22:44.468] ERROR : void Database::populateChildTablesByName(Brewtarget::DBTable) SELECT id FROM fermentable WHERE ( name='Acid Malt'X' AND display=1 ) ORDER BY id ASC LIMIT 1 unrecognized token: "X' AND display=1 ) ORDER BY id ASC LIMIT 1" Unable to execute statement
[20:22:44.468] ERROR : void Database::populateChildTablesByName() void Database::populateChildTablesByName(Brewtarget::DBTable) SELECT id FROM fermentable WHERE ( name='Acid Malt'X' AND display=1 ) ORDER BY id ASC LIMIT 1 unrecognized token: "X' AND display=1 ) ORDER BY id ASC LIMIT 1" Unable to execute statement
[20:22:44.468] ERROR : bool Database::updateSchema(bool*) void Database::populateChildTablesByName(Brewtarget::DBTable) SELECT id FROM fermentable WHERE ( name='Acid Malt'X' AND display=1 ) ORDER BY id ASC LIMIT 1 unrecognized token: "X' AND display=1 ) ORDER BY id ASC LIMIT 1" Unable to execute statement
terminate called after throwing an instance of 'QString'

By changing the SQL statements to escape the name of the record the
issue has been fixed. Fixes issue #319.