Improve handling of fatal exceptions.
While working on issue #319 the application terminated with the following
error message on the console:
terminate called after throwing an instance of 'QString'

Instead of logging this message on the console show a message box to the
user with the contests of the QString thrown. Do the same for
std::exceptions, and a generic message if the exception is of another
type.