Link ingredients to the original parent.
When an ingredient is copied by copying a recipe the newly created
ingredient's parent was the original ingredient. This led to two issues:
* The ingredient could no longer be deleted from the original recipe.
* The inventory of the new ingredient was empty. (Issue #72)

Setting the parent of the newly created ingredient to the parent of the
original ingredient fixes both issue.

Note: since I'm not too familiar with the database structure, please
review carefully; I might have missed something.