Refactoring MainWindow's constructor
I've reduced the almost 400 line MainWindow constructor to 66 lines (including
comments), mostly by moving the obvious groupings into other private methods.
I find this a lot easier to search and read.

I've also moved a few of the signals to the newer method pointer syntax,
because I think the syntax is cleaner and easier to read. It also provides
compile time checking instead of runtime. And that is always better. This
required removing the setRecipe(QModelIndex) method, but that's no great loss.
It was only used in two places, and both of those can handle null pointers on
their own.