Possibly closes #166 and #158
This is a hard thing.

The essence is that in BtLineEdit::setText(BeerXMLElement *element), the
conversion went from QVariant to QString to double. The problem was that the
QString ended up as something like "1.045". The conversion from QString to
double is locale aware, so en_US would get 1.045 but locales like bt_PR would
get 1e3+45.

I fixed this by replacing the QVariant->QString->double conversion with a
QVariant->double conversion. It *seems* to work -- I've tried this with en_US,
bt_PR, de_DE and fr_FR and they all *appear* to do the right thing.

I don't think we need to fix anything else. The values written to the database
are written as doubles, so this kind of conversion is irrelevant. On the other
hand, I have demonstrated an annoying inability to test my own code and my
database my not behave like anybody elses.