Fix warning during compile of Algorithms.h
Polynomials call intPow(), which takes a short unsigned int, not a size_t. order() should therefore also be an unsigned int, and the cast from size_t should be done explicitly at the time of assignment. This also gets rid of the warning that pops up several times about an implicit cast when compiling (at least on VS2015).