Closes #357 in version 2.3.1
There was an off-by-one error that caused MashWizard to reference a newly created
MashStep when determining the temperature before the sparge. This meant that
it always assumed the mash was at the default temperature (67 C) before sparging,
and caused the sparge water temperature to be the same regardless of the actual
mash temperature prior to sparging.