Removed the Debian /usr/local warning.
The CMake code seem to really not like having the CMAKE_INSTALL_PREFIX
as anything else than "/usr/" on Debian but as far as I can see, there
is no reason for it.

Obviously that must have been done for a reason in 2010, but it's not
obvious to me. Furthermore, the Filesystem Hierarchy Standard
specifically mentions that sys admin should install software in
/usr/local and not /usr, the last being only for read-only files.

Furthermore, I think a newer practice is to install apps in /opt/
instead.

The goal of this push request is mainly to stop CMake shooting at me
everytime because I want to install in the standard location.