Step one -- new methods defined, signals working
This is the first commit attempting to clean up the database work again.

The long goal is that we will get all the info from the db in one query
and pass that to the constructors. They will set the cache values only.

Setters will set cache, then set the db values. That will cause the
signal to fire to update everything that needs updating.

Getters will only ever read from cache.

This is not going to be easy, but I think it will be useful.