Fermentables now mostly does what I want
Database::populateElements now runs one query and sends the results to
the Fermentable contructor. The cache values are built and no more. I
can now enable the forwardOnly flag and have it work.

The exception is the inventory tables. I need to really examine how that
works and determine the proper way of handling it. Right now it is
commented out.