brewnotes now follow the new pattern
This will be slow and marginally annoying.

I did refactor the setNotes() method. It was only ever called from one
place and notify was always false. So I removed the flag.

I also added the Q_PROPERTY for attenuation. I know we needed that for
something, but hang me if I can remember why.