Cleaning up mistakes
I had to redo a lot of brewnote.cpp -- the loading flag was really
important and the logic had to be redone in the calculated setters.
Someday, I need to clean that up.

I removed the save() signal from fermentables. I am not 100% with this
decision yet, but I need it to work. This also fixes a bug when
importing recipes -- the new fermentables never got saved.

A few other minor adjustments to make things work. Need to test still
with adding new: recipes, fermentables, hops, equipment, instructions
and misc.