Fixed dropping and core dumps in hops
Added a protection against wandering of the edge of the arrays in
converting types to strings. I need to verify this in all the other
places where we do that.

More importantly, I "fixed" added ingredients to recipes. The old way of
doing things meant we created the copy, inserted the key into the hash
and then let the get() populate the element. Since the gets now only
return what was cached, I had to fix this.

It was a very small change in Database::newIngredient to call the new
constructor with the original ingredient. I am arguing if I should look
the newly added record up and use that, but ... this does work.