Recipe fixes
Slowly going through my changes and fixing things. An issue I have
discovered is that we had a lot of signals crashing about, and it sort
of worked by luck.

As an example, consider the recipe name, display and deleted attributes.
We created the recipe, which fired a newRecipeSignal but didn't set
any of the attributes. This would cause the tree to catch that signal,
but see display as false, which would cause the new recipe to not be
shown.

Once I fixed that, I found something similar with the name -- the tree
would get the newRecipeSignal but not have a name. When the name was
set, something went wrong and caused the tree to show an entry with no
name.

I am considering my options but this makes sure adding a new recipe
works.

I also fixed an annoying bug with Unit::fromSI that caused amounts of 0
to get oddly displayed (a float conversion, I suspect)