Styles are more stylish
Styles mostly work with the caching now. I am experimenting with a new
idea on the editors to reduce the number of DB writes we make. It has
the happy side effect of not adding anything to the DB until we save the
style.

On a side note, the style combo box is really annoying. It doesn't sort.
If you add a new style, it goes to the bottom of the list and there is
no way I can find to make that not happen. The  list seems to be sorted
by DB ID.