Renaming and reworking
I was reading that you shouldn't name member variables with a preceding
_ -- it may conflict with builtin and specials. So I fixed all of them.
The two competing ideas are to prefix with m_ (eg, m_name) or a postfix
_ like name_. I haven't decided what I like yet.

I am also in progress on reworking the database interfaces. I really
dislike that every object has to know its properties and column names. I
dislike the DatabaseSchemaHelper class. I dislike that you need to edit
7+ places just to add one column to a table.

This is my idea of how to get around that problem. And so far, it's
working. I can insert entire records into the db with a few loops. I can
remove (maybe) the db column names from the classes. I still need to
work on how to get rid of the XML mappings, but small steps.