Made cacheOnly an member property
And it really, really cleaned up the code. I like this idea much better.

The idea with be that anything that wants to initialize the object
before we flush to the DB will call setCacheOnly(true), do its thing and
then setCacheOnly(false). It is up to the code to flush the object to
DB, but that is a pretty clean code.

I think I can clean almost all of the constant strings out the of
objects now. That would make me happy. The downside is that TableSchema
is gonna get HUGE. Maybe subclass it?