Mashes, MashSteps and a few fixes
Mashes and mashsteps now work with the new paradigm. Possibly even with
the new paraquarters.

I found more shared property names and XML names, so those got moved to
where they needed to be.

The hard part is how linked mashes and mashsteps are. We really do not
create mashsteps without a mash. So I had to rework the mashstep editor
to know the mash. We also are somewhat random in how the editors create
a new element -- in this case, I had to modify MainWindow to work with
the new system.

Also made a few fixes to misc to make sure all of ranges are tested
before we use them. Still don't know what, but something was setting
type values to -1.