Yeast tables working
Ok. Last remaining tables are for recipes and brewnotes. I would like to
dig at brewnotes for a while and see if I can untangle and undo all the
clever in the initialization of a brewnote, so that one is hard.

Once recipe and brewnote are done, I will give water the treatment just
so it works should we ever figure out how to expose that stuff.

Step after that will be to define the relational tables.