brewnotes are done
Brewnotes are a mess. There are two columns in the db that I don't think
are used. I will clean that up as part of this effort.

I renamed all the files, as I suggested. That accounts for a lot of the
thrash on this checkin.

I will need to consider adding a property to the PropertySchema object
to determine if an property is a foreign key (like mash_id on a recipe).
We probably should write foreign keys, but I want them in this format
for when I take a fixing to DatabaseSchemaHelper.

And I do need to define the relationship tables next.