Decided to add a constraints property
Creating the name column is a bit touchy, since we put an extra 'not
null' constraint on it. I had first hard coded it in DatabaseSchema,
then I had special cased it in DatabaseSchema::generateCreateTable().
Neither solution satisfied.

So I added a constraints property to PropertySchema. It cannot be
initialized, only set after the PropertySchema is new'd. This was mostly
to keep the new PropertySchema calls from getting more complex for just
one field. I like this solution better.

I do worry that the trick I used for the column names needs to be used
for everything. If I need it, I will add it later.