PropertySchema redesigned
After realizing I needed to define contraints, it became obvious the
last idea for the PropertySchema class wasn't going to work. The issue
was that any time I found a new database and had to special case
something, I would have had to update PropertySchema to use an array,
etc.

The new approach is (I hope) a lot cleaner. PropertySchema now contains
a QVector of a struct. The vector is sized to as many databases as we
support dynamically. Each slot in the vector corresponds to the
definition of that property for one database. I've some convenience
methods for when the definition is the same for all of them. I think
this works better.

It required a lot of changes. Some for functionality, many for
readability. It compiles cleanly, but I haven't tested functioning yet.
I am tired and want to get this checked in.

I've made a few changes to TableSchema, notably that the table now knows
if it is a base table (eg, style), inventory table, in recipe table or
child table. This should allow us to clean some of the cruft out of
Database and replace it with dynamic queries.

I've modified DatabaseSchema to return all the different kinds of
tables. Next step is to clean house and see how it works.